+++ packages/share_plus/share_plus/android/src/main/kotlin/dev/fluttercommunity/plus/share/MethodCallHandler.kt
@@ -5,7 +5,7 @@ import io.flutter.plugin.common.MethodChannel
 import java.io.IOException
 
 /** Handles the method calls for the plugin.  */
-internal class MethodCallHandler(private val share: Share) : MethodChannel.MethodCallHandler {
+internal class MethodCallHandler(private val share: Share, private val manager: ShareSuccessManager) : MethodChannel.MethodCallHandler {
 
   override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
     when (call.method) {
@@ -15,6 +15,7 @@ internal class MethodCallHandler(private val share: Share) : MethodChannel.Metho
         share.share(
           call.argument<Any>("text") as String,
           call.argument<Any>("subject") as String?,
+          false,
         )
         result.success(null)
       }
@@ -28,8 +29,37 @@ internal class MethodCallHandler(private val share: Share) : MethodChannel.Metho
             call.argument<List<String>?>("mimeTypes"),
             call.argument<String?>("text"),
             call.argument<String?>("subject"),
+            false,
           )
           result.success(null)
+        } catch (e: IOException) {
+          result.error(e.message, null, null)
+        }
+      }
+      "shareWithResult" -> {
+        expectMapArguments(call)
+        if(!manager.setCallback(result)) return
+
+        // Android does not support showing the share sheet at a particular point on screen.
+        share.share(
+          call.argument<Any>("text") as String,
+          call.argument<Any>("subject") as String?,
+          true,
+        )
+      }
+      "shareFilesWithResult" -> {
+        expectMapArguments(call)
+        if(!manager.setCallback(result)) return
+
+        // Android does not support showing the share sheet at a particular point on screen.
+        try {
+          share.shareFiles(
+            call.argument<List<String>>("paths")!!,
+            call.argument<List<String>?>("mimeTypes"),
+            call.argument<String?>("text"),
+            call.argument<String?>("subject"),
+            true,
+          )
         } catch (e: IOException) {
           result.error(e.message, null, null)
         }