+++ packages/share_plus/share_plus/lib/share_plus.dart
@@ -6,6 +6,8 @@ import 'dart:async';
 import 'dart:ui';
 
 import 'package:share_plus_platform_interface/share_plus_platform_interface.dart';
+export 'package:share_plus_platform_interface/share_plus_platform_interface.dart'
+    show ShareResult, ShareResultStatus;
 
 /// Plugin for summoning a platform share sheet.
 class Share {
@@ -72,6 +74,59 @@ class Share {
     assert(paths.isNotEmpty);
     assert(paths.every((element) => element.isNotEmpty));
     return _platform.shareFiles(
+      paths,
+      mimeTypes: mimeTypes,
+      subject: subject,
+      text: text,
+      sharePositionOrigin: sharePositionOrigin,
+    );
+  }
+
+  /// Behaves exactly like [share] while providing feedback on how the user
+  /// interacted with the share-sheet. Until the returned future is completed,
+  /// any other call to any share method that returns a result _might_ result in
+  /// a [PlatformException] (on Android).
+  ///
+  /// Because IOS, Android and macOS provide different feedback on share-sheet interaction,
+  /// a result on IOS will be more specific than on Android or macOS. While IOS can detect if
+  /// the user actually completed his selected action or aborted it midway, Android and macOS
+  /// only record if the user selected an action or outright dismissed the share-sheet.
+  ///
+  /// **Currently only implemented on IOS, Android and macOS.**
+  static Future<ShareResult> shareWithResult(
+    String text, {
+    String? subject,
+    Rect? sharePositionOrigin,
+  }) async {
+    assert(text.isNotEmpty);
+    return _platform.shareWithResult(
+      text,
+      subject: subject,
+      sharePositionOrigin: sharePositionOrigin,
+    );
+  }
+
+  /// Behaves exactly like [shareFiles] while providing feedback on how the user
+  /// interacted with the share-sheet. Until the returned future is completed,
+  /// any other call to any share method that returns a result _might_ result in
+  /// a [PlatformException] (on Android).
+  ///
+  /// Because IOS, Android and macOS provide different feedback on share-sheet interaction,
+  /// a result on IOS will be more specific than on Android or macOS. While IOS can detect if
+  /// the user actually completed his selected action or aborted it midway, Android and macOS
+  /// only record if the user selected an action or outright dismissed the share-sheet.
+  ///
+  /// **Currently only implemented on IOS, Android and macOS.**
+  static Future<ShareResult> shareFilesWithResult(
+    List<String> paths, {
+    List<String>? mimeTypes,
+    String? subject,
+    String? text,
+    Rect? sharePositionOrigin,
+  }) async {
+    assert(paths.isNotEmpty);
+    assert(paths.every((element) => element.isNotEmpty));
+    return _platform.shareFilesWithResult(
       paths,
       mimeTypes: mimeTypes,
       subject: subject,