+++ packages/share_plus/share_plus_macos/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.1.0
+
+- Add *WithResult methods to get feedback on user action
+
 ## 2.0.2
 
 - Obtain Flutter view from the registrar