+++ packages/share_plus/share_plus_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.1.0
+
+- Add *WithResult methods to get feedback on user action
+
 ## 2.0.1
 
 - Improve documentation